#18631029 : fixed the case that update and install operation shows repeatedly in easyinstall module when updated with FTP
git-svn-id: http://xe-core.googlecode.com/svn/sandbox@7214 201d5d3c-b55e-5fd7-737f-ddc643e51545

+++ modules/autoinstall/autoinstall.class.php
@@ -46,11 +46,13 @@
          **/
         function checkUpdate() {
             $oDB =& DB::getInstance();
-            if($oDB->isTableExists("autoinstall_installed_packages"))
+            if(!file_exists(FileHandler::getRealPath("./modules/autoinstall/schemas/autoinstall_installed_packages.xml"))  
+                && $oDB->isTableExists("autoinstall_installed_packages"))
             {
                 return true;
             }
-            if($oDB->isTableExists("autoinstall_remote_categories"))
+            if(!file_exists(FileHandler::getRealPath("./modules/autoinstall/schemas/autoinstall_remote_categories.xml"))  
+                && $oDB->isTableExists("autoinstall_remote_categories"))
             {
                 return true;
             }
@@ -63,11 +65,13 @@
          **/
         function moduleUpdate() {
             $oDB =& DB::getInstance();
-            if($oDB->isTableExists("autoinstall_installed_packages"))
+            if(!file_exists(FileHandler::getRealPath("./modules/autoinstall/schemas/autoinstall_installed_packages.xml"))  
+                && $oDB->isTableExists("autoinstall_installed_packages"))
             {
                 $oDB->dropTable("autoinstall_installed_packages");
             }
-            if($oDB->isTableExists("autoinstall_remote_categories"))
+            if(!file_exists(FileHandler::getRealPath("./modules/autoinstall/schemas/autoinstall_remote_categories.xml"))  
+                && $oDB->isTableExists("autoinstall_remote_categories"))
             {
                 $oDB->dropTable("autoinstall_remote_categories");
             }
